+++ src/main/java/com/udi/gaaf/metodo_pago/MetodoPagoService.java
@@ -61,6 +61,11 @@ public class MetodoPagoService {
 	
 	public DatosDetalleResponse eliminarPorId(Long id) {
 		var metodo = obtenerMetodoPagoPorId(id);
+		
+		if(metodo.getMedio() != null) {
+			throw new BadRequestException("No se puede eliminar hay asociado medio de pagos con ese metodo");
+		}
+		
 		repository.delete(metodo);
 		return new DatosDetalleResponse(200, "Metodo de pago eliminado correctamente");
 	}
